Add ibm_db2 e2e test (#4411)
