+++ ibm_db2/tox.ini
@@ -7,6 +7,8 @@ envlist =
     bench
 
 [testenv]
+description =
+    py{27,37}: e2e ready
 dd_check_style = true
 usedevelop = true
 platform = linux|darwin|win32
